podman: drop workaround to skip podman-tests when not running on CKI

+++ container/podman/runtest.sh
@@ -98,10 +98,12 @@ function run_tests()
         rstrnt-report-result "${RSTRNT_TASKNAME}/info" FAIL
     fi
 
-    # return before executing podman-tests
-    echo "WARN: not executing podman tests due to lack of test maintainer capacity"
-    echo "more details: https://gitlab.com/redhat/centos-stream/tests/kernel/kernel-tests/-/issues/1502"
-    return $TEST_FAILED
+    if [[ -n "$CKI_NAME" ]]; then
+        # return before executing podman-tests if it is a CKI job
+        echo "WARN: not executing podman tests due to lack of test maintainer capacity"
+        echo "more details: https://gitlab.com/redhat/centos-stream/tests/kernel/kernel-tests/-/issues/1502"
+        return $TEST_FAILED
+    fi
 
     # Clear images
     run_cmd_user "podman system prune --all --force && podman rmi --all"
